+++ src/index.js
@@ -2,4 +2,4 @@ import './js/api';
 import './js/popular-films-fetch';
 import './js/teem-modal';
 import './js/fetch-one-film';
-import './js/traler';
+import './js/fetchVideoKey';
