+++ src/js/fetch-one-film.js
@@ -1,6 +1,10 @@
 import axios from 'axios';
 import Notiflix from 'notiflix';
-import defaultImage from '/src/images/no-poster.png'
+import defaultImage from '/src/images/no-poster.png';
+//for Trailer
+import * as basicLightbox from 'basiclightbox';
+import 'basiclightbox/dist/basicLightbox.min.css';
+import fetchVideoKey from './fetchVideoKey';
 
 const body = document.querySelector('body');
 const listMovies = document.querySelector('.js-cards-list');
@@ -26,11 +30,11 @@ backdropMovie.addEventListener('click', event => {
 });
 
 //-----функції------
-
+let idMovie = undefined;
 async function onOpenModalMovieClick(event) {
   // отримує id конкретного фільму
   const perentNodLi = event.target.closest('li');
-  const idMovie = perentNodLi?.dataset?.id;
+  idMovie = perentNodLi?.dataset?.id;
   // перевірка що користувач клікнув саме на картку фільму
   if (idMovie === undefined) {
     return;
@@ -50,6 +54,9 @@ async function onOpenModalMovieClick(event) {
   renderModalMovieInfo(movieInfo);
   // ховає спінер
   Notiflix.Loading.remove();
+  // слухач для трейлера
+  const iframeRef = document.querySelector('.js-iframe');
+  iframeRef.addEventListener('click', onClickYouTube);
 }
 
 // закриття модалки фільму
@@ -136,7 +143,7 @@ function renderModalMovieInfo(movieInfo) {
     alt="${title}"
     data-id="${id}"
     />
-  <a class="link-trailer">Trailer</a>
+  <a class="link-trailer js-iframe">Trailer</a>
 </div>
 <div class="modal-card__thumb-right">
   <p class="thumb-right__title">${title}</p>
@@ -174,4 +181,30 @@ function renderModalMovieInfo(movieInfo) {
 </div>`;
 
   cardMovie.insertAdjacentHTML('beforeend', markup);
+}
+
+// Trailer
+
+function onClickYouTube() {
+  fetchVideoKey(idMovie).then(key => {
+    const instance = basicLightbox.create(
+      `
+		<iframe width="560" height="315" src="https://www.youtube.com/embed/${key}" frameborder="0" allowfullscreen></iframe>
+	`,
+      {
+        onClose: () => {
+          window.removeEventListener('keydown', onEscKeyDown);
+        },
+      }
+    );
+    instance.show();
+
+    window.addEventListener('keydown', onEscKeyDown);
+    function onEscKeyDown(event) {
+      if (event.code === 'Escape') {
+        instance.close();
+      }
+      console.log(event);
+    }
+  });
 }
