+++ src/partials/header-home.html
@@ -20,9 +20,8 @@
           </li>
         </ul>
         <ul class="authorization">
-          <li class="authorization__js-signIn">
-            <!-- authorization__btn -->
-            <button class="iframe">
+          <li class="authorization__js-signIn">і
+            <button class="authorization__btn">
               <svg width="16" height="16">
                 <use href="/src/images/icons.svg#icon-sign-in"></use>
               </svg>
